export run method
